backend/requirements.txt
django==4.2.1
mysqlclient==2.2.7
python-dotenv
python-dotenv==1.1.0
pillow==11.0.0
django-storages==1.13.2
boto3==1.33.1
drf-yasg==1.21.10
djangorestframework==3.16.0
django-minio-storage==0.5.8
django-cors-headers==4.3.1
channels>=4.0
requests
faker
